+++ android/src/main/kotlin/com/hyiso/flutter/plugins/hybrid_flutter/HybridFlutterManager.kt
@@ -1,4 +1,4 @@
-package io.hybrid.flutter.plugins.hybrid_flutter
+package com.hyiso.flutter.plugins.hybrid_flutter
 
 import android.content.Context
 import io.flutter.embedding.android.FlutterFragment