+++ android/src/main/kotlin/com/hyiso/flutter/plugins/hybrid_flutter/HybridFlutterPlugin.kt
@@ -1,4 +1,4 @@
-package io.hybrid.flutter.plugins.hybrid_flutter
+package com.hyiso.flutter.plugins.hybrid_flutter
 
 import io.flutter.Log
 import io.flutter.embedding.engine.FlutterEngine