+++ example/android/app/src/main/kotlin/io/hybrid/flutter/plugins/hybrid_flutter_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.hybrid.flutter.plugins.hybrid_flutter_example
+package com.hyiso.flutter.plugins.hybrid_flutter_example
 
 import android.app.Activity
 import android.content.Intent