+++ example/android/app/src/main/kotlin/io/hybrid/flutter/plugins/hybrid_flutter_example/MainFlutterFragment.kt
@@ -1,8 +1,8 @@
-package io.hybrid.flutter.plugins.hybrid_flutter_example
+package com.hyiso.flutter.plugins.hybrid_flutter_example
 
 import android.net.Uri
 import io.flutter.embedding.android.FlutterFragment
-import io.hybrid.flutter.plugins.hybrid_flutter.HybridFlutterFragment
+import com.hyiso.flutter.plugins.hybrid_flutter.HybridFlutterFragment
 
 class MainFlutterFragment : HybridFlutterFragment() {